Moved Anki2Context into the Data Access Layer.
