+++ AnkiSentenceCardBuilder/AnkiJapaneseFlashcardManager.csproj
@@ -51,7 +51,6 @@
     <Folder Include="ApplicationLayer\DTOs\Config\" />
     <Folder Include="ApplicationLayer\Interfaces\Services\Managements\" />
     <Folder Include="ApplicationLayer\Services\Managements\" />
-    <Folder Include="DataAccessLayer\Contexts\" />
     <Folder Include="DataAccessLayer\Interfaces\Contexts\" />
     <Folder Include="DataAccessLayer\Repositories\" />
     <Folder Include="DomainLayer\Entities\" />
