+++ AnkiSentenceCardBuilder/Services/UpdateAnki2.cs
@@ -1,7 +1,7 @@
 using System.IO;
 using System.Threading.Tasks;
+using AnkiJapaneseFlashcardManager.DataAccessLayer.Contexts;
 using AnkiSentenceCardBuilder.Controllers;
-using AnkiSentenceCardBuilder.Models;
 using Azure.Storage.Blobs.Specialized;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
